@@ -26,8 +26,8 @@
     "docs:build": "vitepress build docs",
     "docs:preview": "vitepress preview docs",
     "format": "prettier \"**/*\" --ignore-unknown --write",
-    "lint": "eslint --ext .js,.ts,.mjs,.mts,.vue ./",
-    "type-check": "vue-tsc --noEmit -p tsconfig.vue-tsc.json",
+    "lint": "eslint -c ./eslint.config.js './**/*.{js,ts,cjs,mjs,mts,vue}'",
+    "type-check": "vue-tsc --noEmit -p tsconfig.json",
     "test": "echo \"No test specified\" && exit 0"
   },
   "dependencies": {
@@ -69,8 +69,8 @@
   "devDependencies": {
     "@daipeng7/rollup-plugin-iconfont": "^3.0.6",
     "@electron/rebuild": "^3.7.1",
-    "@intlify/vite-plugin-vue-i18n": "^7.0.0",
-    "@quasar/app-vite": "^1.11.0",
+    "@intlify/unplugin-vue-i18n": "^6.0.0",
+    "@quasar/app-vite": "^2.0.0",
     "@rollup/plugin-inject": "^5.0.5",
     "@sentry/esbuild-plugin": "^2.22.7",
     "@sentry/vite-plugin": "^2.22.7",
@@ -82,15 +82,15 @@
     "@types/node": "^20.17.8",
     "@types/pretty-bytes": "^5.2.0",
     "@types/sanitize-html": "^2.13.0",
-    "@typescript-eslint/eslint-plugin": "^7.18.0",
-    "@typescript-eslint/parser": "^7.18.0",
+    "@vue/eslint-config-prettier": "^10.1.0",
+    "@vue/eslint-config-typescript": "^14.1.3",
     "autoprefixer": "^10.4.20",
     "electron": "^33.2.0",
     "electron-builder": "^25.1.8",
-    "eslint": "^8.57.1",
-    "eslint-config-prettier": "^9.1.0",
+    "eslint": "^9.16.0",
     "eslint-plugin-perfectionist": "^4.2.0",
     "eslint-plugin-vue": "^9.32.0",
+    "globals": "^15.13.0",
     "husky": "^9.1.7",
     "lint-staged": "^15.2.10",
     "postcss": "^8.4.49",
@@ -104,7 +104,7 @@
   ],
   "lint-staged": {
     "*": "prettier --write --ignore-unknown",
-    "*.{js,ts,mjs,mts,vue}": "eslint --fix"
+    "*.{js,ts,cjs,mjs,mts,vue}": "eslint --fix"
   },
   "engines": {
     "node": "^20.18.0",
@@ -116,6 +116,7 @@
     "win32"
   ],
   "private": true,
+  "type": "module",
   "packageManager": "yarn@4.5.3",
   "productName": "Meeting Media Manager"
 }
