@@ -1,4 +1,4 @@
-import { boot } from 'quasar/wrappers';
+import { defineBoot } from '#q-app/wrappers';
 import messages from 'src/i18n';
 import { createI18n } from 'vue-i18n';
 
@@ -7,7 +7,7 @@ export type MessageSchema = (typeof messages)['en'];
 
 let i18n: ReturnType<typeof createI18n> = createI18n({});
 
-export default boot(({ app }) => {
+export default defineBoot(({ app }) => {
   i18n = createI18n({
     fallbackLocale: 'en',
     legacy: false,
