@@ -1,17 +1,17 @@
+import { defineBoot } from '#q-app/wrappers';
 import { init, replayIntegration } from '@sentry/electron/renderer';
 import {
   browserTracingIntegration,
   init as initVue,
   vueIntegration,
 } from '@sentry/vue';
-import { boot } from 'quasar/wrappers';
 import { IS_DEV } from 'src/constants/general';
 import { SENTRY_DSN } from 'src/constants/sentry';
 import { errorCatcher } from 'src/helpers/error-catcher';
 
 import { version } from '../../package.json';
 
-export default boot(({ app, router }) => {
+export default defineBoot(({ app, router }) => {
   try {
     if (!IS_DEV) {
       init(
