@@ -120,7 +120,7 @@
               <q-icon color="primary" :name="icon" />
             </q-item-section>
             <q-item-section>
-              <q-item-label>{{ $t(name) }}</q-item-label>
+              <q-item-label>{{ $t(name ? name : '') }}</q-item-label>
               <q-item-label caption>{{ $t(name + '-explain') }}</q-item-label>
             </q-item-section>
           </q-item>
@@ -347,6 +347,7 @@ const minDate = () => {
     const dateArray: Date[] =
       lookupPeriod.value[currentCongregation.value]?.map((day) => day.date) ||
       [];
+    if (!dateArray[0]) return undefined;
     const minDate = getMinDate(dateArray[0], ...dateArray.slice(1));
     return formatDate(minDate, 'YYYY/MM');
   } catch (error) {
@@ -361,6 +362,7 @@ const maxDate = () => {
     const dateArray: Date[] =
       lookupPeriod.value[currentCongregation.value]?.map((day) => day.date) ||
       [];
+    if (!dateArray[0]) return undefined;
     const maxDate = getMaxDate(dateArray[0], ...dateArray.slice(1));
     return formatDate(maxDate, 'YYYY/MM');
   } catch (error) {
@@ -381,6 +383,7 @@ const dateOptions = (lookupDate: string) => {
     const dateArray: Date[] =
       lookupPeriod.value[currentCongregation.value]?.map((day) => day.date) ||
       [];
+    if (!dateArray[0]) return true;
     const minDate = getMinDate(dateArray[0], ...dateArray.slice(1));
     const maxDate = getMaxDate(dateArray[0], ...dateArray.slice(1));
     return (
