@@ -129,7 +129,7 @@ const populatePublicTalks = async () => {
     s34mpDb.value,
     'SELECT DISTINCT VersionNumber, Year FROM Publication',
   );
-  if (PublicationInfos.length) s34mpInfo.value = PublicationInfos[0];
+  if (PublicationInfos[0]) s34mpInfo.value = PublicationInfos[0];
 };
 
 const browse = async () => {
@@ -137,8 +137,10 @@ const browse = async () => {
   if (!s34mpFileSelection || !s34mpFileSelection.filePaths.length) return;
   s34mpFile.value = s34mpFileSelection.filePaths[0];
   if (s34mpDir.value) await fs.ensureDir(s34mpDir.value);
-  await decompressJwpub(s34mpFile.value, s34mpDir.value, true);
-  populatePublicTalks();
+  if (s34mpFile.value) {
+    await decompressJwpub(s34mpFile.value, s34mpDir.value, true);
+    populatePublicTalks();
+  }
 };
 
 const dismissPopup = () => {
