@@ -85,7 +85,8 @@ const exportDayToFolder = async (targetDate?: Date) => {
   const sections: Partial<Record<MediaSection, number>> = {}; // Object to store dynamic section prefixes
   for (let i = 0; i < dayMediaLength; i++) {
     try {
-      const m = dynamicMediaFiltered[i];
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      const m = dynamicMediaFiltered[i]!;
       const sourceFilePath = window.electronApi.fileUrlToPath(m.fileUrl);
       if (
         !sourceFilePath ||
