@@ -1,5 +1,5 @@
+import type { JsonObject } from 'app/node_modules/obs-websocket-js/node_modules/type-fest';
 import type { OBSWebSocketError } from 'obs-websocket-js';
-import type { JsonObject } from 'obs-websocket-js/node_modules/type-fest';
 import type { ObsConnectionState, ObsSceneType } from 'src/types';
 
 import { defineStore } from 'pinia';
